# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PROBOX Documentación",
        "description": "Adminstración de Proyectos APP",
        "contact": {
            "email": "daruiza@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8090/api",
            "description": "PROBOX API Server"
        }
    ],
    "paths": {
...
        "/evidence/showtaskbyid/{id}": {
            "get": {
                "tags": [
                    "Evidence"
                ],
                "summary": "Get One Task By one Id",
                "description": "Return One Task by Id",
                "operationId": "getShowTaskById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Show Task by Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/project/index": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Get All Project",
                "description": "Return Project",
                "operationId": "getProject",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/project/store": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Store Project",
                "description": "Store Project",
                "operationId": "storeProject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Project"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/project/showbyid/{id}": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Get One Project By one Id",
                "description": "Return One Project",
                "operationId": "getProjectById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/project/update/{id}": {
            "put": {
                "tags": [
                    "Project"
                ],
                "summary": "Update One Project By one Id",
                "description": "Update One Project",
                "operationId": "getUpdateProjectById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Project"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/project/destroy/{id}": {
            "delete": {
                "tags": [
                    "Project"
                ],
                "summary": "Delete One Project By one Id",
                "description": "Delete One Project",
                "operationId": "getDestroyProjectById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/project/showtaskbyid/{id}": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Get Tasks By one Id",
                "description": "Return Tasks by Id",
                "operationId": "getShowTaskById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Show Tasks by Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Project": {
                "title": "Project",
                "description": "Project body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the Module",
                        "type": "string",
                        "example": "Name Module"
                    },
                    "price": {
                        "title": "price",
                        "description": "Price of the Project",
                        "example": "Price Project"
                    },
                    "date_init": {
                        "title": "date_init",
                        "description": "Date_init of the Project",
                        "type": "string",
                        "format": "date",
                        "example": "Date_init Project"
                    },
                    "date_closed": {
                        "title": "date_closed",
                        "description": "Date_closed of the Project",
                        "type": "string",
                        "format": "date",
                        "example": "Date_closed Project"
                    },
                    "address": {
                        "title": "address",
                        "description": "Address of the Project",
                        "type": "string",
                        "example": "Address Project"
                    },
                    "quotion": {
                        "title": "quotion",
                        "description": "Quotion of the Project",
                        "type": "string",
                        "example": "Quotion Project"
                    },
                    "goal": {
                        "title": "goal",
                        "description": "Goal of the Project",
                        "type": "string",
                        "example": "Goal Project"
                    },
                    "photo": {
                        "title": "photo",
                        "description": "photo of the Project",
                        "type": "string",
                        "example": "photo Project"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the Project",
                        "type": "string",
                        "example": "Description Project"
                    },
                    "focus": {
                        "title": "focus",
                        "description": "Focus of the Project",
                        "type": "boolean",
                        "example": "Focus Project"
                    },
                    "active": {
                        "title": "active",
                        "description": "Active of the Project",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}